+++ i386/include/apm_bios.h
@@ -12,7 +12,7 @@
  *
  * Aug, 1994	Implemented on FreeBSD 1.1.5.1R (Toshiba AVS001WD)
  *
- * $FreeBSD: src/sys/i386/include/apm_bios.h,v 1.32 2005/04/03 23:26:05 imp Exp $
+ * $FreeBSD: src/sys/i386/include/apm_bios.h,v 1.33 2009/09/27 14:00:16 jkim Exp $
  */
 
 #ifndef	_MACHINE_APM_BIOS_H_
@@ -230,12 +230,12 @@ typedef struct apm_pwstatus {
 } *apm_pwstatus_t;
 
 struct apm_bios_arg {
-        u_long eax;
-        u_long ebx;
-        u_long ecx;
-        u_long edx;
-        u_long esi;
-        u_long edi;
+        uint32_t eax;
+        uint32_t ebx;
+        uint32_t ecx;
+        uint32_t edx;
+        uint32_t esi;
+        uint32_t edi;
 };
 
 struct apm_event_info {
